added a guard to make sure app does not crash when deleting alarms

+++ app/src/main/java/se/jakob/knarkklocka/data/MainAlarmViewModel.java
@@ -23,6 +23,8 @@ public class MainAlarmViewModel extends AndroidViewModel {
     }
 
     public void delete() {
-        mRepository.delete(mAlarm.getValue());
+        if (mAlarm.getValue() != null) {
+            mRepository.delete(mAlarm.getValue());
+        }
     }
 }
